use TaskSubOperation for group MS2 task
